Database: changed the datatype of average_rating column of recipe table from Integer into Float

+++ model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, String, JSON, ForeignKey
+from sqlalchemy import Integer, Float, String, JSON, ForeignKey
 from sqlalchemy.orm import mapped_column
 from config import db
 
@@ -24,7 +24,7 @@ class Recipe(db.Model):
     image = mapped_column(String(255), nullable=False)
     datetime = mapped_column(String(100), nullable=False)
     user_id = mapped_column(Integer, ForeignKey("user.id"))
-    average_rating = mapped_column(Integer, nullable=False, default=0)
+    average_rating = mapped_column(Float, nullable=False, default=0)
     total_ratings = mapped_column(Integer, nullable=False, default=0)
 
 class Rating(db.Model):
